Mark the repository objects to be able to identify them even after they have been rewrapped. An update (update_334_335) is needed.
Simplify non-indexing of repository objects.

When the proxy tool is notified about a modification of a repository
object, in addition to reindexing the proxies for the object it also
sends a notification of modification for them. This is important to get
a Tree Cache notified of a change of the Title, for example.