This module deals with trees, and how they are modified using a set of simple add/modify/delete operations.
Its goal is to get a minimal representation on how to get from one tree
to another, by simplifying redundant operations. We're only interested
in the state as it exists in the final tree. This is useful to lazily
replay changes to a master tree, without doing useless operations.

Nodes in the tree are addressed using a sequence, usually a tuple, which
represents a path.

The operations available are ADD, REMOVE, MODIFY.

This will be used for the tree caches, to allow delayed updates.